requirements.txt
flask
flask-sqlalchemy
sqlalchemy
flask-migrate
flask-minify
alembic
flask-wtf
wtforms
flask-login
bs4
lxml
pyyaml
markupsafe
pillow
ipaddress
numpy
flask-httpauth
requests
unidecode
email_validator
selenium
chromedriver
celery
redis
flask-socketio
flask-babel
pybluemonday
docxtpl
packaging
gostcrypto
